Convert MaskRuleConfiguration to DistSQL via SPI
